+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Nothing yet!
+### Fixed
+
+- Fix issues with dev servers using Svelte 5 with the Vite plugin ([#15250](https://github.com/tailwindlabs/tailwindcss/issues/15250))
 
 ## [4.0.0-beta.4] - 2024-11-29